feat: track duration int -> long으로 변경

+++ core/domain/src/main/java/com/hsikkk/delightroom/domain/model/valueobject/MediaPlayerStatus.kt
@@ -17,7 +17,7 @@ data class MediaPlayerStatus(
 ){
     val currentTrack: Track?
         get() = playList.getOrNull(currentTrackIndex)
-    val duration: Int?
+    val duration: Long?
         get() = currentTrack?.duration
 }
